continued LED refactoring / memory optimization
finally some success!
device objects are still located in memory at the beginning of .data -- out of reach from The Evil Stackpointer.
LEDs are now objects for mj808 and 818 -- 828 is a harder nut since it involves the charlieplexing handler and a ton of flags/variables.
LED refactoring still continues.

the move to .data was done in the previous commit, yet now it is verified to work with example LED objects.

command parsing was moved to PopulatedBusOperation(), where it sort of belongs more than before.

code structure changes:
- changed declaration and definition of device objects (.h and .c files): objects are now in .h extern and defined into .data in the corresponding .c files

in mj8x8.c and .h:
- changed signature of PopulatedBusOperation()
- removed digest() function implementation

in mj808.c and .h and similar device-specific files:
- defined device-specific enumerations

in led.c and .h:
- reorganized led structs -- still WIP

in main.c:
- in INT1_vect() modified call to PopulatedBusOperation()
- in TIMER1_COMPA_vect() started to access LEDs via their object