continued information hiding
finally got rid of the u_devices union in canbus_t and replaced it with a normal uint16_t value called __devices.
using bitmasks and #defined's in mj8x8.h __devices can be queried for whatever is or is not on the bus.
for public access a getter called GetDevices() was introduced into __message_handler_t

in mj828.c and .h:
- refactored  charlieplexing_handler(composite_led_t *in_led) to charlieplexing_handler(uint8_t in_flags) as a preparation to privatize ledflags_t

in main.c:
- adapted code to work with above changes