finished HR30_6R_6PD(71) - FFC layout.brd & .sch this item allows for FFC connectivity between the base board and external connectors. the FFC has a pitch-out of 1mm (FFC width of 11mm) and mates with a HR30_6R_6PD(71) TH receptacle and a Samtecc FC1-10-02-T-WT FFC connector.
finished MJ838 - base - rev.1a.brd
this item is an improved version of the original base board.
it now can accept either the RA or straight version of the MEC1-150-02 card edge connector.
this board can be connected via wires or via the above HR30_6R_6PD(71) - FFC layout.brd.
for wired connection terminal blocks are used, for FFC connection a Samtec FC1-10-02-T-WT is used.
!!! these rev.1 base boards are NOT COMPATIBLE with earlier boards !!!

finished MJ838 - rev.1c.brd & .sch
this item is the next revision of the MJ838 device.
it follows a dual-device design started with rev.1b - the schematic/board are developed to support both mj838 and mj000 devices, depending on which firmware is loaded and which components are populated.
major improvements for mj838/mj000 in respect to rev.1b:
- far more powerful MCU (STM32L4P5CGU6) - more flash/ram, floating point, SDMMC, QSPI
- improved zerocross detection via TLV6703DDCR's internal refernece voltage
- addition of numerous TVS protection elements (SWD, ZC & phase signals, SDMMC interface)
- addition of the ADXL367 - Accelerometer for improved motion detection
- addition of QSPI for MRAM storage
- switch from SPI to SDMMC for interfacing SD Card storage
- addition of RN4678 bluetooth module (via UART) for wireless data transfer
- improved routing of MJ8x8 power bus, application load lines to card edge (and into base board) so that prod<->dev card swap is far easier
- addition of NFET controlled application/adjustable load switch

added old boards for reference; their name prefix is "reference - *"